Learn to find servers using krb5.conf and DNS
Use ctypes to load libkrb5.so.3 and ask it where the KDCs are, and if
DNS lookups are allowed (we default to True if there's no explicit
configuration).  If there are no configured servers, and DNS lookups are
allowed, search DNS for a realm's servers.  Try all addresses we get
back until we succeed in connecting to one.