Handle cases where the target-realm isn't set
Do some deeper parsing to handle cases where the target-realm isn't set
in the encapsulating proxy message.  Pull the target realm name out of
the KDC-REQ or AP-REQ, and if we have a target-realm, make sure that
they match.  If we don't have a target-realm, just use the name that we
dug out of the inner request.